pimodfeed/csgo-src

+++ vgui2/matsys_controls/manipulator.cpp
@@ -1,4 +1,4 @@
-//========= Copyright © 1996-2001, Valve LLC, All rights reserved. ============
+//========= Copyright Valve Corporation, All rights reserved. ============//
 //
 // Purpose: 
 //
@@ -19,10 +19,6 @@
 #include "mathlib/mathlib.h"
 
 #include <float.h>
-
-// NOTE: This has to be the last file included!
-#include "tier0/memdbgon.h"
-
 
 using namespace vgui;
 
@@ -146,7 +142,7 @@ void CPotteryWheelManip::OnCursorMoved( int x, int y )
 	{
 		m_azimuth  += 0.002f * ( x - m_lastx );
 		m_altitude -= 0.002f * ( y - m_lasty );
-		m_altitude = MAX( -M_PI/2, MIN( M_PI/2, m_altitude ) );
+		m_altitude = MAX( (float)-M_PI/2, MIN( (float)M_PI/2, m_altitude ) );
 	}
 	m_lastx = x;
 	m_lasty = y;
